+++ app/src/main/java/com/ssafy/ssaign/src/splash/SplashActivity.kt
@@ -1,22 +1,33 @@
 package com.ssafy.ssaign.src.splash
 
+import android.Manifest
 import android.content.Intent
+import android.content.pm.PackageManager
+import android.os.Build
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
+import androidx.annotation.RequiresApi
 import com.ssafy.ssaign.config.BaseActivity
 import com.ssafy.ssaign.databinding.ActivitySplashBinding
 import com.ssafy.ssaign.src.main.MainActivity
+import com.ssafy.ssaign.src.main.permission.PermissionActivity
 
 class SplashActivity : BaseActivity<ActivitySplashBinding>(ActivitySplashBinding::inflate) {
     private val DELAY = 500 // 원하는 값으로 설정
 
+    @RequiresApi(Build.VERSION_CODES.M)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         Handler(Looper.getMainLooper()).postDelayed({
-            startActivity(Intent(this, MainActivity::class.java))
-            finish()
+            if(checkSelfPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
+                startActivity(Intent(this, MainActivity::class.java))
+                finish()
+            } else {
+                startActivity(Intent(this, PermissionActivity::class.java))
+                finish()
+            }
         }, 500)
     }
 }
