+++ app/src/main/res/layout/fragment_permission_denied.xml
@@ -1,0 +1,64 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:tools="http://schemas.android.com/tools"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    tools:context=".src.main.permission.PermissionDeniedFragment">
+
+    <androidx.constraintlayout.widget.Guideline
+        android:id="@+id/guideline1"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal"
+        app:layout_constraintGuide_percent="0.32" />
+
+    <TextView
+        android:id="@+id/fragment_permission_denied_tvTitle"
+        style="@style/header_text"
+        android:text="@string/permission_denied_title"
+        android:gravity="center"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="@+id/guideline1" />
+
+    <View
+        android:id="@+id/line"
+        android:layout_width="match_parent"
+        android:layout_height="0.5dp"
+        android:background="@color/black_low_emphasis"
+        android:layout_margin="32dp"
+        app:layout_constraintTop_toBottomOf="@id/fragment_permission_denied_tvTitle"/>
+
+    <TextView
+        android:id="@+id/fragment_permission_denied_tvTitle2"
+        style="@style/body_text"
+        android:text="@string/permission_denied_title2"
+        android:textStyle="bold"
+        android:layout_marginTop="32dp"
+        app:layout_constraintTop_toBottomOf="@+id/line"
+        app:layout_constraintStart_toStartOf="@id/line"
+        app:layout_constraintEnd_toEndOf="@id/line"/>
+
+    <TextView
+        android:id="@+id/fragment_permission_denied_tvContent"
+        style="@style/explain_text"
+        android:gravity="center"
+        android:layout_marginTop="16dp"
+        android:text="@string/permission_denied_content"
+        app:layout_constraintTop_toBottomOf="@+id/fragment_permission_denied_tvTitle2"
+        app:layout_constraintStart_toStartOf="@id/line"
+        app:layout_constraintEnd_toEndOf="@id/line"/>
+
+    <Button
+        android:id="@+id/fragment_permission_denied_btnBack"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="이전"
+        android:textSize="@dimen/header_text_size_16sp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
+
